LayoutTests/fast/css/shorthand-mismatched-list-crash-expected.txt

Test for WebKit bug 31559: Crash with mismatched lists and shorthands.

FAIL para.style.webkitTransition should be width 1s, left 1s, 1s. Was width 1s ease 0s, left 1s ease 0s, 1s ease 0s.
FAIL para.style.webkitTransition should be width 1s, left 1s, top. Was width 1s ease 0s, left 1s ease 0s, top.
PASS successfullyParsed is true

TEST COMPLETE
